# Speed up quad_segment_maxima by skipping unneeded lineparam calculation.